Change locking order for retrying jobs
To attach ci_builds as a partition to p_ci_builds we need
to lock ci_builds, ci_stages, and ci_pipelines in access
exclusive mode. This change makes sure that we're locking
them in the correct order in the application code
to reduce the number of places from which it could deadlock.