Implement redis store for circuitbox
By default, the circuit breaker is using a memory store to track
the state of the circuit. This is okay for a single node
environment since threads can share the store.

On production though, we have multiple nodes (puma and sidekiq) so
a need for a persisted store is required.

This adds a custom store that will utilize redis as the storage. It
uses the redis-ratelimiting DB for this since it's kind of related
with rate limiting.

The `Gitlab::CircuitBreaker::Store` is based on Circuitbox's memory
store (https://github.com/yammer/circuitbox/blob/main/lib/circuitbox/memory_store.rb)
but modified to use redis.

This is behind the `circuit_breaker` feature flag.