Ensure protected branches audit event even if cache refresh fails
We had an issue where a protected branch was successfully destroyed,
yet no audit event was present.

We were not able to reproduce the issue locally, as the happy path
successfully results in an audit event being logged.

It was noticed that, if a cache refresh was to fail, we could end up
in this state where the protected branch was deleted and no audit
event was logged.

This change rescues runtime errors around the cache refresh to prevent
this hypothetical from occurring