Define different service names per LLM client
`Gitlab::Llm::Concerns::CircuitBreaker` requires `service_name` to
be defined.

Before this change, we are only using a single `service_name` and
that means all client will use a single circuit. If a single provider
fails and the circuit opens, all providers will be affected.

To prevent that, since we have different clients (e.g. OpenAI, Vertex,
Anthropic), we define a specific service name per client.

This also includes a fix to `ExponentialBackoff` concern to raise
the correct exception to avoid a `NameError`.