blueprints: Detail object network-based architecture for object pools
Add documentation on the current object pool-based architecture and add
an overview of the new object network-based architecture that raises the
abstraction level and moves the ownership of object pool lifecycle
management into Gitaly. Furthermore, add a rough migration plan that
spells out the different steps we need to perform in order to arrive at
the new architecture.