Lower ProcessSyncEventsWorker dedupe TTL to 1 minute
Previously the TTL was set to the default of 6 hours, which caused
delays in syncing of the `Ci::NamespaceMirror` entries. This sync is
crucial for users to have access to CI-related resources, such as
runners.

Lower this TTL for both projects and namespaces to 1 minute. Over the
past week, the longest job took 38 seconds, and that only happened
once.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/409089

Changelog: changed