Add a global utility for message I/O in the Code Review area
The purpose of this is to separate generic things like
 "socket messages" and "observable streams" out from
 being connected to "a component."

Sometimes, it might make sense to connect a socket
directly to a component's data or to a data store.
In those cases, the `apollo` key of a Vue makes the
most sense. However, sockets and observables are useful
outside of just setting a property or updating data.
In this case, we're attaching to a *signal*. Signals
can be used in powerful applications like modifying
a state machine, or in simple applications like how
this is used, where we just call a method.

Because we've chosen Apollo as our tool for
communicating across Websockets, there's quite a bit of
overhead associated with this signaling system, but
by abstracting signal IO to a central place not in any
particular component, we can re-use the code anywhere
in the application.