Add gitleaks allowlist for secret-detection
We have recurring secrets being reported by Secret Detection. These
secrets are only used for tests, and are not real tokens, but they keep
appearing in Vulnerability Reports.

This MR is introducing a gitleaks config with an allowlist so these
secrets are not reported anymore, once for all.

Using `#gitleaks:allow` comments could work too, but sometimes secrets
are in the docs, and adding a comment there on the same line isn't
possible.