Add version to sentry events
This change publishes the current version of GitLab along with events
sent to Sentry.

Currently both revision and release show the same sha information, this
change removes the duplication and sets release as the proper version of
Sentry.

The legacy integration of Sentry that still currently enabled by default
is not affected, so no changelog is added.