Add test to validate route-to-primary gRPC metadata
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/119765
introduced a regression that caused GitLab Rails to drop the route-to
-primary gRPC metadata, which caused any changes pushed to Gitaly
Cluster repositories to fail intermittently:
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/14259.

To ensure this behavior is maintained, we add a GRPC interceptor in
the `/api/v4/internal/base` request test and ensure ensuing gRPC
requests have this metadata. Note that we can only test this with
project and snippet repositories because the Git change access checks
appear to invoke gRPC calls, while Wiki repositories don't appear to
invoke any gRPC calls.