Return blank object_url for audit events that have a deleted entity
This was an error occuring in some QA specs.

I think it may only occur due to the way audit_events and their
associations are created/manipulated in specs. I do not think this
occurs in the normal application code flow

Either way, it makes sense to return a blank `object_url` from the
presenter of the entity object is unavailable for some reason

resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/408156