Update URL scheme for packages hosted on PyPI
GitLab makes hyperlinks for Python packages listed in `requirements.txt`
to PyPI, the Python Package Index, a public service by the Python
Foundation.

Those packages used to be hosted on http://pypi.python.org, but the
hostname (and webserver implementation) changed a few years ago. PyPI is
now available at https://pypi.org, packages are referred to with the
`/project/` path, and the URL path always has a trailing slash.