User Profile - Snippets in Vue Init
This change aims to being the
conversion the User Profile
Snippets tab to Vue.

The goal here is to try and
keep the UI as similar as
possible to the current
implementation.

This change in particular
focuses on setting up the tab
and pagination.  The snippet
rows will be added in a follow
up MR.