Remove mentions of performance problems
This page is purely a FYI for customers/and or policy for our support
team, but can be read easily internally as engineering guidelines.

Text like "can cause performance and stability issues" are warnings
rather than permitting us to release experimental/beta features that
are un-performant.

This edit removes these passages as they could easily be misinterpreted
by others internally and lead to us referencing this as policy so we
merge something if its experimental/beta even if a reviewer has
identified performance issues to be addressed.