doc/praefect: Replication factor recommendations
The default replication factor replicates repositories to every storage
in the Gitaly Cluster deployment. This is problematic for large
deployments because replica repositories put additional pressure on the
primary repository during replication events.

This change updates replication factor documentation to recommend
setting the configuration explicitly instead of relying on the default
for large deployments.