Implement FillInMergeRequestTemplate AI action
Adds support for `fillInMergeRequestTemplate` input for `aiAction`
that has the following arguments:
- `resourceId` (required) - project where the MR will be created on
- `title` (required) - title of the MR to be created
- `sourceProjectId` (optional, defaults to `resource`)
- `sourceBranch` (required)
- `targetBranch` (required)
- `content` (required)

The `sourceBranch` and `targetBranch` will be used to compare
the changes between those branches so we can get the diff to send
to the AI provider. That's why it's required.

The `sourceProjectId` is needed for forks so we can compare the
changes made on the fork's source branch with the target branch.

Sections in the template content can add `<!-- AI Skip -->` to
the body to ensure that AI won't touch the section.

This uses Vertex AI as the AI provider.

This is behind the `fill_in_mr_template` feature flag.