Proper handle null values returned from BitBucket Cloud API
In some situations BitBucket Cloud API can return a null author. For
this case we should map the original pull request author to an anonymous
user.

Also when a pull request is declined, BitBucket Cloud API the pull
request does not have the target commit SHA as the pull request was
not merged

Changelog: fixed