Add logging to MergeRequests::AfterCreateService
To help investigate https://gitlab.com/gitlab-com/ops-sub-department/section-ops-request-for-help/-/issues/122
further, this adds some logging calls in the service to determine
where it stops when it gets killed.

Based on recent investigation, we know that the NewMergeRequestWorker
gets killed (but not all the time) for a certain project. My theory
is that the project has integrations (it's already confirmed they
have integrations) and generating the hook data can get expensive.

The logging calls here can help confirm if the worker gets killed
while trying to execute the hooks before it tries to create a
pipeline.