Fix 500 error for domain verification
When SSO is enforced and a top-level group owner attempts to add
a domain for verification, a 500 error can occur if the user
does not have an active SSO session. This change will cause the
user to be redirected to SSO sign-in in the absence of an SSO
session.

Changelog: fixed
EE: true