Fix validation in UI when updating notes
- Check if note is valid in the Notes::UpdateService.
Return the note with errors it if isn't.

- Return errors as json from the #update method in NotesController

- Display the validation error in UI and
restore note content when updating fails

- Update deprecated_notes to do the same, since snippets
and commits still use deprecated_notes to render comments

- Fix specs for noteable_note.vue.
Since noteable_note.vue mutates the note prop that gets passed down to
it, it was causing strange side effects in the specs.
Hence, create a clone of the note object in createWrapper() method
and pass that down as the prop in each test

Changelog: fixed