Fix LSN Host#caught_up?/replica_is_up_to_date? for logical replicas
These methods previously only relied on `pg_last_wal_replay_lsn()` for
replicas. In a logical replica this function cannot be meaningfully
compared against the LSN from the primary because they have forked. As
such we need to make use of
`remote_lsn from pg_replication_origin_status` which provides a way to
determine where the logical replication is up to with respect to LSNs
from the primary. This solution was proposed in
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23578#note_1396755188
.