Remove single tab from admin mode
In the admin mode, the single tab felt redundant because it showed the
same title as the breadcrumb, which was very close. In the redesign of
the sign-in pages the single tab is hidden. Likewise, we do the same in
the admin mode. This brings consistency between the main sign-in page
and the admin mode.

Changelog: changed