Work around macOS forking issues with grpc gem update
With the update of the grpc gem to v1.55.0, it appears that after a
grpc connection is completed, the status is parsed, and the timestamp
is formatted with the local time zone. The underlying C++ library,
abseil-cpp, appears to call Apple system APIs to retrieve the local
time zone. However, macOS will crash the Puma process if these system
APIs are called within a fork.

We currently have a workaround for macOS
(https://bugs.ruby-lang.org/issues/14009) that prevents the crash for
the Foundation framework. However, not only does the the grpc update
need the CoreFoundation framework, but it also needs to call
`CFTimeZoneCopyDefault` to load the default time zone. We now use the
`ffi` gem to load the default time zone and release the resource to
avoid this crash, as reported in
https://github.com/grpc/grpc/issues/26257.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/413536

Changelog: fixed