Revert ReleasesResolver and add VersionsResolver
ReleasesResolver was recently updated to use BatchLoader
which returns a null value when a project does not have
releases. This caused the FE project releases page to break
since it expects an empty array. This MR reverts the changes
to ReleasesResolver and creates a new VersionsResolver
to takes its place for the CiCatalogResource.versions
field.

Changelog: fixed