Return 503 instead of 429 when Gitaly returns ResourceExhausted
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/119054, GitLab
Rails propagates Gitaly's ResourceExhausted response codes back to
users. It returns a friendly message and a 429 status code. This HTTP
status turns out to be confusing. It's about the user sending too many
requests, aka the user's responsibility to back off. Unfortunately, it's
not the case here.

Gitaly returns ResourceExhausted response codes when it wants to push
back traffic as it's at capacity. Gitaly applies different limiters,
mostly concurrency limiting, not rate limiting. As a result, 429 does
not reflect the system's actual behavior.

This commit switches the returned HTTP code to 503. This status code is
appropriate for the described case. The Retry-After header stays intact,
as it's a supported use case:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Retry-After

Changelog: changed