Fix memory leak in CI config includes entry
In GitLab 11.9,
https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/24098
introduced validation of `include` keywords. However, it quietly
introduced a memory leak: any time a new `include` entry was
instantiated, a proc would be added to the `aspects` class variable.
As the list grew, the time taken to process other `include` keywords
would grow.

This commit fixes this by doing away with the class variable in favor of
`ComposableArray`.

Changelog: fixed