[Audit Events] Provide all audit events under a group via the API
Background Context
------------------------

We want to be able to return all audit events under a given groups via
a single endpoint. We have an issue, [_API endpoint to retrieve all
audit events under a given group_][0], to track this feature.

We actually had already implemented that feature, but the performance
was so poor in production that it functionally did not work. The
database query timed out almost 100% of the time.

The proposal was to improve the existing query's performance by using
the [`InOperatorOptimization`][1] library.

Attempts to do this did not show a good performance boost (details
[here][2])

Another Approach
----------------

We then thought about implementing this feature with a new
approach. The Idea is that we can add a new column, `group_rollup`, to
the `audit_events` table. This column will be the `id` of the
top-level group the audit event belongs to.

We could then update the Auditor to write this value at audit event
creation, allowing us to implement the endpoint with a simple query:

```sql
SELECT *
FROM "audit_events"
WHERE "audit_events"."group_rollup_id" == @group.id
```

This query performed much better in database lab (details in this
[spike issue][3])

Back to InOperatorOptimization
---------------------------------
In the MR for the `group_rollup_id` approach, it was discovered that
with some tweaking and adding a new index, the InOperatorOptimization
might just work

So this commit adds the application code for that approach.

The index additions are tracked in these issues:

- [asynchronous database index addition][4]
- [synchronous database index addition][5]

[0]:https://gitlab.com/gitlab-org/gitlab/-/issues/337757
[1]:https://docs.gitlab.com/ee/development/database/efficient_in_operator_queries.html
[2]:https://gitlab.com/gitlab-org/gitlab/-/issues/337757#note_1312002908
[3]:https://gitlab.com/gitlab-org/gitlab/-/issues/396778#note_1325762680
[4]:https://gitlab.com/gitlab-org/gitlab/-/issues/411167
[5]:https://gitlab.com/gitlab-org/gitlab/-/issues/411129

Changelog: changed
EE: true