Add default database values for user_preferences table
We replaced default_value_for gem with the native rails implementation
in f124341c18ed43e63a0e2184ae9fcfc6d5978184.
We added some setters and getters to ensure the same behavior. After
adding the default for columns, we need to backfill the null values
and then remove the setters and getters which is a bit hacky way to
make this work.

Changelog: other