Fix Spring not working with macOS
Previously the macOS-specific initializer was only run in Puma's
`before_fork` block, but this code only gets executed when Puma is in
clustered mode. However, Spring also needs this initializer because it
works by forking the application.

We can simplify the initializer by always running it in macOS.
This works because:

1. Spring loads the application first, running all the initializers,
and then forking the process.
2. Puma preloads the application via `preload_app!` so the
initializer runs before the process forks.
3. Sidekiq and Rake tasks don't fork, but it doesn't hurt to run the
initializer.

In addition, it appears that we need to add a call to
`CFTimeZoneGetName` since otherwise we get a `[__NSCFConstantString
initialize] may have been in progress in another thread when fork()
was called.` error under Spring. This was discovered by attaching a
debugger to the forked process with
`GRPC_CONFIG=dbg gem install grpc -v 1.55.0`.

Changelog: fixed