Remove Finding.by_state check_feedback argument
With the deprecate_vulnerabilities_feedback now on by default the
Vulnerability::Feedback model should no longer be used.

This change removes the `check_feedback` argument from the
`Security::Finding.by_state` scope.  This flag causes the scope
to add an `OR` condition on the query pulling in from the
`vulnerability_feedback` table when true.

Once this is removed there are no remaining callers of the scope
`Security::Finding.dismissed_by_feedback` so that scope is removed here
too.