Fix false positives for Migration/SchemaAdditionMethodsNoPost cop
We allow `add_column` and `create_table` methods inside the `down`
method in post-deploy migrations. But if these methods were wrapped
inside a statement like `if` then the cop was raising an error.
For example, in the following, the 2nd example was being incorrectly
flagged as an offense:

```
def down
  add_column(:table, :column)
end
```

```
def down
  add_column(:table, :column) unless column_exists(:table, :column)
end
```