Combine sign-in text together - part 1/2
There are three different options that change the appearance of the
sign-in pages (some of the options below affects the sign-up and
trial_registration pages too):

1. Admin > Settings > Appearance > Sign in/Sign up pages > Description
1. Admin > Settings > Preferences > Sign-in and Help page > Additional
   text to show on the sign-in page
1. Admin > Settings > General > Sign-in restriction > Sign-in text

In this MR we revert the order of fields to before the redesign of the
sign-in pages.

We also remove the `Need help` title and icon.

In part 2 we will combine the three settings into a single option.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/375290
Closes https://gitlab.com/gitlab-org/gitlab/-/issues/410889

Changelog: changed