Add a blueprint for Gitaly Adaptive Concurrency Limit
Gitaly, a Git server, needs to push back on its clients to reduce the
risk of incidents. In the past, we introduced per-RPC concurrency limit
and pack-objects concurrency limit. Both systems were successful, but
the configurations were static, leading to some serious drawbacks. This
blueprint proposed an adaptive concurrency limit system to overcome the
drawbacks of static limits. The algorithm primarily uses the Additive
Increase/Multiplicative Decrease approach to gradually increase the
limit during normal processing but quickly reduce it during an issue.
The algorithm focuses on lack of resources and serious latency
degradation as criteria for determining when Gitaly is in trouble.