Ensure a group is always left with at least one owner
When we remove an owner from a group, we need to be sure that there is
a remaining owner. This checked failed in case the remaining owner was
a deactivated user: in that case, the last owner could be removed and
the group was left without any owner.

Changelog: changed

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/356909