Add BigInt `dependencyProxyTotalSizeBytes` field
This MR deprecates the `dependencyProxyTotalSizeInBytes` 
field and introduces the `dependencyProxyTotalSizeBytes` 
field. The new field is of type `BigInt`, which returns 
a string instead of a number.

The dependency_proxy_total_size_in_bytes is the sum of all of the the group’s
dependency_proxy_manifests and dependency_proxy_blobs size in bytes.
And both of those tables have a size column of type bigint.

When the customer has a total size larger than 2147483647, an IntegerEncodingError
is thrown with 'Integer out of bounds'.

Changelog: added