Fix rendering of imported file comments
- Fix rendering of discussions that do not have "diff_file.diff_refs".
Some comments on a whole file (imported from GitHub)
did not render properly as they use LegacyDiffNotes on the
backend which does not include "diff_refs". This fix adds a
fallback value for the old / new SHA to use "commit_id"
which when passed, renders the correct file path.
- Add "no_preview" option as a diffFile viewer option. This makes
sure we don't show the "replaced" side-by-side option for
file comments.

Changelog: fixed