Add Add-on purchase API services
New API endpoints to sync Add-on purchases are going to be added in
another iteration. These endpoints allow to create, get and update an
Add-on purchase. In this change, two new service to create and update an
Add-on purchase are added that will later be used by the endpoints.

Changelog: added
EE: true