Drop Sidekiq jobs based on feature flag
This adds new feature to DeferJobs middleware to drop the job entirely
instead of deferring them. This is useful during an incident where we
know we don't want to resume the job execution, whereas deferring them
piles up the ScheduledSet and potentially leads to thunderring herd
problem once the jobs are released.

Changelog: added