Makes slow feature spec faster (2x faster)
For this given spec, we were generating a huge dataset to persist
into the Rails cache, and to check whether we would render cached
data (when the dataset isn't present in the cache, we display an
error message to the user).

The change introduced in this commit is to generate a fake dataset to be
persisted in the cache, as we don't need to have the real dataset: we
just want it to be present in the cache, and we'll return that dataset.