Update CI/CD catalog details page to use API data
We now fetch most of the data for the CI/CD catalog
details page through the API instead of using a client
resolver. This requires som small tweaks in the
component to use the new data structure and update
tests. This is behind a disabled FF,
`ci_namespace_catalog_experimental`