Stop using allow_disk_access in wiki specs
These wiki specs were using raw filesystem access to determine what HEAD
was set to. This will not be possible through praefect since it will not
be possible to determine where the replicas are located. So instead we
use the new `head_only` option to fetch HEAD over a RPC.