Remove setup instructions from suggestions docs
This removes the setup instructions for the GitLab VS Code extension from the
documenation. The VS Code docs cover the setup instructions and should
be the authoritative source for how to configure things.

This resolves the difference that the code suggestions docs
proposed different API scopes than what the extension actually
required.