Remove 5s replication lag from qa-with-load-balancing
This was proving to be too flaky and I was unsure if this was worth it
from the beginning. Even without the replication lag we're still
validating that GitLab works correctly with multiple databases
configured and there is still some very small replication lag that we
can expect whenever we use Postgres replication.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/415648