Load gems from all groups in preflight server boot
We used to only use the gem production group, but review-apps
apparently try to load gems from all groups.

The purpose of the preflight checks are to catch as many errors
as possible, as early as possible. Loading all gems by default
when booting the application server is a pretty good strategy
to catch errors related to gems.

It's arguable whether the review-apps actually need to load
non-production gems, as it's supposed to mirror production
gitlab deployments.