Do not broadcast AI responses twice
When the Agent picks the `SummarizeComments` tool, we are internally
calling the `GenerateSummaryService`, which also stores the response to
the cache and broadcasts it via the GraphQL subscription.

With `skip_cache: true` we were already able to not store the response
to the cache. This change renames the `skip_cache` option to
`internal_request` and also no longer broadcasts the response to the
client, which resulted in duplicated responses in the chat.