Fix empty response for #refs_by_oid method
Sentry error:
https://new-sentry.gitlab.net/organizations/gitlab/issues/221174

**Problem**

We expect `#refs_by_oid` to return a collection of references. However,
it can also return `nil`. That leads to `NoMethodError` in cases when
developer forgets to check for `nil` value.

**Solution**

Return an empty array instead of `nil`. It will prevent accidental
mistakes and streamline the code.

Changelog: fixed