Reduce GitlabSettings::Options interface
Instead of delegating any Hash method to GitlabSettings::Options, we're
exposing a smaller subset of Hash methods to avoid unexpected behavior.

For backward compatibility, if a not allowed Hash method is called in
production we still delegate to GitlabSettings::Options internal hash,
but we raise an exception in Development and Test environment.