Remove feature flag "find_changed_paths_new_format"
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/414463

Use new GRPC format to request changed paths for commits.
We explicitly provide a list of commit requests instead of an array of
commit hashes.

This changee is backward compatible with a previous implementation, but
it does not rely on deprecated `commits` field.

Changelog: other