Fail after 20 specs failure by default
Failed specs can take a long time to run, due to potential timeouts
before reaching those errors. This currently makes some RSpec jobs
run for a long time, and this lengthen the pipelines duration.

This commit introduces a limit to the number of failures before
we fail the RSpec job that ran them (defaults to 20 for now)