Sync both default_branch_protection columns on API update
We are migrating `default_branch_protection` from an `integer` column to a
`jsonb` column.

- The original column is `default_branch_protection`
- The new column is `default_branch_protection_defaults`

Right now, we have only added the new `jsonb` column. It is not yet
exposed via the API and not yet being used.

In this commit, we are just syncing the existing column to the new one
on API updates.

This is meant to accomplish a couple of things:

1. reduce the number of rows that will eventually need to be
   backfilled for the new column
2. allow us to eventually test some of the app code without exposing
   the new column in the API yet