Disable `rake gitlab:cleanup:remote_upload_files` with bucket prefix
In GitLab 15.0
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/91307 added
official support for configuring an object storage bucket with a
prefix. However, this Rake task doesn't take this bucket prefix into
account and attempts to iterate through all files in the bucket. If
the dry-run flag is disabled, this Rake task also moves all files into
the lost and found directory.

Unfortunately, it does not appear Fog provides an easy, cloud-agnostic
way to list all files in a bucket with a prefix filter. In addition,
at least for Azure Blob Storage, there isn't a standardized method to
distinguish a directory from a regular file using Fog.

For these reasons, this commit disables this Rake task if a prefix is
configured to prevent data loss.

This Rake task should probably be dropped for a number of reasons:

1. It's not used very much.
2. It requires bucket permissions to list all files. Our documented
    permissions for object storage buckets don't grant these
    privileges.
3. It requires walking through the entire bucket and doing a database
    query for each batch size. This is quite slow, and it doesn't
    scale well as more objects are added.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/415537

Changelog: changed