Do not retry if we have the max number of failures
If we have RSPEC_FAIL_FAST_THRESHOLD tests failures, we might
have more tests failures happening.

If we only retry the RSPEC_FAIL_FAST_THRESHOLD tests that failed,
we might miss other failures, and pass the CI job,
which is not intended.

This commit ensures that we do not auto-retry the failed tests
when we have the maximum of allowed failures.