Replace .sr-only with aria-label
While the screen-reader-only label seems redundant given the link
already has accessible text, it exists so that the confirmation modal
picks up the text for its confirm button.

For some reason, [it explicitly looks][1] for screen-reader-only text.
The [original MR][2] does not explain why. There is an [issue][3] to
explore changing this.

Why switch from .sr-only to aria-label:
- To be consistent with link button above it.
- The `sr-only` class is from Bootstrap, and generally we're trying to
  move away from that.

[1]: https://gitlab.com/gitlab-org/gitlab/-/blob/59409bf02df839aabb2cfee70975458b82260f68/app/assets/javascripts/lib/utils/confirm_via_gl_modal/confirm_via_gl_modal.js#L6-9
[2]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/73167
[3]: https://gitlab.com/gitlab-org/gitlab/-/issues/413033