Refactor runner actions
This internal change splits the buttons component logic for pausing
and deleting individual runners. So the resulting components can be
reused in using a different appearance.

Splits mutation button components into delete action (with the mutation)
and button (presentation).

For example, the delete action can now be used in presentations that
are not buttons.