Log Gitaly metadata when a Gitaly call raises an exception
Recently, when a call to Gitaly raises an exception, that exception's
message, class, and backtrace. Unfortunately, it doesn't tell anything
about the target node. That lack of information makes it hard to debug
when something goes wrong. Although we can follow the user flow via
correlation ID, the case that Gitaly node is not available is
overlooked.

This commit adds Gitaly metadata to exception structured logs.

Changelog: added