Remove UUID calculation workaround for namespace_id
We knew that Rails' UUID v5 calculation was not correct according to
the RFC, therefore, we had to implement a workaround to calculate the
UUIDs correctly.

With Rails 7, they fixed this issue and put the new fix behind a config
parameter to prevent breaking the UUIDs of existing applications.

With this commit, we remove our workaround and set that configuration
parameter to stop getting deprecation warnings.