objectpools: Add `pool_repositories` table index
Currently the relative path of an object pool uniquely identifies the
corresponding repository because an individual object pool cannot exist
on multiple storages. In the future, repository storage moves may result
in the same object pool being replicated to another storage. To support
this, the unique index of the `pool_repositories` table is updated to
use both the `disk_path` (relative path) and `shard_id` (storage).

Changelog: added